// typescript/user-interface/packages/ui-app/src/ts/components/analyst-ui/components/waveform/types.ts
export interface WaveformDisplayReduxProps {
  // passed in from golden-layout
  glContainer?: GoldenLayout.Container;
  currentTimeInterval: CommonTypes.TimeRange;
  currentStageName: string;
  defaultSignalDetectionPhase: CommonTypes.PhaseType;
  currentOpenEventId: string;
  selectedSdIds: string[];
  stationsVisibility: AnalystWaveformTypes.StationVisibilityChangesDictionary;
  selectedStationIds: string[];
  selectedSortType: AnalystWorkspaceTypes.WaveformSortType;
  analysisMode: WorkflowTypes.AnalysisMode;
  measurementMode: AnalystWorkspaceTypes.MeasurementMode;
  sdIdsToShowFk: string[];
  location: AnalystWorkspaceTypes.LocationSolutionState;
  channelFilters: Record<string, FilterTypes.Filter>;
  openEventId: string;
  keyPressActionQueue: Record<string, number>;
  // because the user may load more waveform
  // data than the currently opened time interval
  viewableInterval: WeavessTypes.TimeRange;
  zoomInterval: WeavessTypes.TimeRange;
  minimumOffset: number;
  maximumOffset: number;
  baseStationTime: number;
  shouldShowTimeUncertainty: boolean;
  shouldShowPredictedPhases: boolean;
  qcMaskQuery: { data: any };
  qcSegments: Record<string, Record<string, QcSegment>>;
  processingMask: ChannelSegmentTypes.ProcessingMask;
  maskVisibility: Record<string, boolean>;
  alignablePhases: CommonTypes.PhaseType[];
  stationsAssociatedWithCurrentOpenEvent: string[];
  phaseToAlignOn: CommonTypes.PhaseType | undefined;
  alignWaveformsOn: AlignWaveformsOn;
  waveformClientState: WaveformLoadingState;
  filterList: FilterTypes.FilterList;

  // callbacks
  isStationVisible(station: StationTypes.Station | string): boolean;
  isStationExpanded(station: StationTypes.Station | string): boolean;
  getVisibleStationsFromStationList(stations: StationTypes.Station[]): StationTypes.Station[];
  pan(
    panDirection: WaveformTypes.PanType,
    options?: { shouldLoadAdditionalData?: boolean; onPanningBoundaryReached?: () => void }
  ): WeavessTypes.TimeRange;
  setDefaultSignalDetectionPhase(phase: CommonTypes.PhaseType): void;
  setMode(mode: AnalystWorkspaceTypes.WaveformDisplayMode): void;
  setOpenEventId(eventId: string): void;
  setSelectedSdIds(idx: string[]): void;
  setSelectedStationIds(ids: string[]);
  setSdIdsToShowFk(signalDetections: string[]): void;
  setSelectedSortType(selectedSortType: AnalystWorkspaceTypes.WaveformSortType): void;
  setChannelFilters(filters: Record<string, WaveformTypes.WaveformFilter>);
  setMeasurementModeEntries(entries: Record<string, boolean>): void;
  setKeyPressActionQueue(actions: Record<string, number>): void;
  setStationsVisibility(
    stationsVisibility: AnalystWaveformTypes.StationVisibilityChangesDictionary
  );
  setStationVisibility(station: StationTypes.Station | string, isVisible: boolean): void;
  setStationExpanded(station: StationTypes.Station | string, isExpanded?: boolean): void;
  setChannelVisibility(
    station: StationTypes.Station | string,
    channel: ChannelTypes.Channel | string,
    isVisible: boolean
  ): void;
  setViewableInterval(viewableInterval: CommonTypes.TimeRange): void;
  setMinimumOffset(minimumOffset: number): void;
  setMaximumOffset(maximumOffset: number): void;
  setBaseStationTime(baseStationTime: number): void;
  setStationsAssociatedWithCurrentOpenEvent(stationsAssociatedWithCurrentOpenEvent: string[]): void;
  setZoomInterval(zoomInterval: CommonTypes.TimeRange): void;
  showAllChannels(station: StationTypes.Station | string): void;
  setShouldShowTimeUncertainty(newValue: boolean): void;
  setShouldShowPredictedPhases(newValue: boolean): void;
  markAmplitudeMeasurementReviewed(args: any): Promise<void>;
  onWeavessMount?(weavessInstance: WeavessTypes.WeavessInstance): void;
  setAlignWaveformsOn(alignWaveformsOn: AlignWaveformsOn): void;
  setPhaseToAlignOn(phaseToAlignOn: CommonTypes.PhaseType): void;
}

/**
 * Consolidated props type for waveform display.
 */
export type WaveformDisplayProps = WaveformDisplayReduxProps &
  ProcessingAnalystConfigurationQueryProps &
  PredictFeaturesForEventLocationQueryProps &
  StationQueryProps & {
    events: EventTypes.Event[];
    signalDetections: SignalDetectionTypes.SignalDetection[];
    channelSegments: Record<string, Record<string, UiChannelSegment[]>>;
    uiTheme: ConfigurationTypes.UITheme;
    eventStatuses: Record<string, EventStatus>;
    distances: EventTypes.LocationDistance[];
    offsets: Record<string, number>;
  };

/**
 * The props for the {@link WaveformComponent}.
 * We omit the signalDetectionResults and channelSegmentResults and replace them with
 * the modified fetch results type, because the non-ideal state component consumes
 * and removes the metadata (such as isLoading and isError). This is a performance
 * optimization, since it reduces the number of times the {@link WaveformPanel} renders
 */
export type WaveformComponentProps = Omit<
  WaveformDisplayProps,
  'channelSegments' | 'events' | 'signalDetections'
>;